Renamed to OnSegment.h
