+++ content/geometry/SegmentIntersection.h
@@ -22,7 +22,7 @@ Products of three coordinates are used in intermediate steps so watch out for ov
 #pragma once
 
 #include "Point.h"
-#include "onSegment.h"
+#include "OnSegment.h"
 
 template<class P> vector<P> segInter(P a, P b, P c, P d) {
 	auto oa = c.cross(d, a), ob = c.cross(d, b),
