+++ content/geometry/chapter.tex
@@ -8,7 +8,7 @@
 	\kactlimport{SegmentIntersection.h}
 	\kactlimport{lineIntersection.h}
 	\kactlimport{sideOf.h}
-	\kactlimport{onSegment.h}
+	\kactlimport{OnSegment.h}
 	\kactlimport{linearTransformation.h}
 	\kactlimport{Angle.h}
